22107.201_beta6
The final Beta! After so many trial and errors...

+++ core/main.ps1
@@ -12,86 +12,9 @@ function Show-Branding($s1) {
 $workdir = "$PSScriptRoot\.."
 
 ##############################################################
-### Everything inside these two lines of 63 hashes are the 
-### only thing you should modify. PLEASE DO NOT MOFIDY 
-### ANYTHING ELSE IN THIS FILE!
+# Load variables from the configuration file
+. $PSScriptRoot\config.ps1
 
-### ------ SCRIPT CONFIGURATION: Module Switches ------
-
-	$setupmusic =           $true
-
-if ($pwsh -eq 7) {
-	$windowsupdate =        $true}
-if ($pwsh -eq 5) {
-	$dotnet462 =            $true}
-	$dotnet35 =             $true
-	$essentialapps =     	$true
-	$hidetaskbaricons =     $true
-	$removeedgeshortcut =   $true
-	$setwallpaper =         $true
-	$desktopshortcuts =     $true 
-	$removeudpassistant =   $true 
-	$removewaketimers =     $true 
-	$removeUWPapps =        $true 
-	$openshellconfig =      $true # Requires $essentialapps
-	$explorericon =         $true
-	$taskbarpins =          $true
-	$replaceemojifont =     $true 
-	$defaultapps =          $true 
-	$disablesettings =      $true 
-	$blockapps =            $true 
-	$gpomodifs =            $true 
-	$disableaddressbar =    $true
-	$removeonedrive =       $true
-	$removehomegroup =      $true
-	$explorerstartfldr =    $true 
-	$oldbatteryflyout =     $true 
-	$customsounds =         $true
-	$removesystemapps =     $true
-	$sltoshutdownwall =		$true
-
-### ------ SCRIPT CONFIGURATION: Registry Switches ------
-# Below are registry-applied tweaks. You can disable all of them,
-# or you can disable a specific one.
-
-	$registrytweaks =       $true # Master switch for all option below
-	
-	$disablestartupitems =  $true
-	$disablenotifs =        $true 
-	$disablegamebar =       $true 
-	$disableautoplay =      $true 
-	$disablemultitaskbar =  $true
-	$transparencydisable =  $true 
-	$disableanimations =    $true 
-	$disablewinink =        $true 
-	$removedownloads =      $true
-	$applookupinstore =     $true
-	$contextmenuentries =   $true
-	$1709peoplebar =        $true
-	$1511locationicon =     $true
-	$removequickaccess =    $true
-	$disablelocationicon =  $true 
-	$activatephotoviewer =  $true 
-	$registeredowner =      $true 
-	$classicpaint =         $true 
-	$disableedgeprelaunch = $true
-	$disablecortana =       $true 
-	$disablestoreautoupd =  $true
-	$balloonnotifs =        $true 
-	$showalltrayicons =     $false # Recommended TRUE only if 1511+
-	$disablelockscrn =      $true
-	$darkmodeoff =          $true 
-	$classicalttab =        $true
-	$oldvolcontrol =        $true
-	$defaultcolor =         $true
-	$remove3Dobjects =      $true
-	$hidebluetoothicon =    $true
-	$disablelogonbg =       $true
-	$removelckscrneticon =  $true
-	$svchostslimming =      $true
-	$desktopversion =       $true
-	
-##############################################################
 # Importing some basic functions required for the modules menu
 function Set-WallPaper($Image) {
 Add-Type -TypeDefinition @" 
@@ -117,7 +40,7 @@ public class Params
 function Get-ConfigStat {
 	$confuled = (Get-ItemProperty -Path "HKCU:\Software\AutoIDKU").ConfigSet
 	if ($confuled -eq 2) {
-		Set-ItemProperty -Path "HKCU:\Software\AutoIDKU" -Name "RebootScript" -Value 1 -Type DWord -Force
+		Set-ItemProperty -Path "HKCU:\Software\AutoIDKU" -Name "RebootScript" -Value 3 -Type DWord -Force
 		exit
 	}
 	elseif ($confuled -eq 0) {
@@ -159,27 +82,41 @@ if ($hkm -eq 1) {
 # Continue importing required dependencies
 Write-Host -ForegroundColor Cyan -BackgroundColor DarkGray "Initializing components" -n; Write-Host ([char]0xA0)
 
-switch ($build) {
-    {$_ -ge 10240} {
-        $pendingreboot = (Test-Path 'HKLM:\SOFTWARE\Microsoft\Windows\CurrentVersion\Component Based Servicing\RebootPending')
-        if ($pendingreboot -eq $true -and $build -ge 10240) {
-            Write-Host -ForegroundColor Cyan -BackgroundColor DarkGray "Your PC has a pending restart, which has to be done before running this script. Automatically restarting in 5 seconds..." -n; Write-Host ([char]0xA0)
-			Set-ItemProperty -Path "HKCU:\Software\AutoIDKU" -Name "HikaruMode" -Value 1 -Type DWord -Force
-			& $PSScriptRoot\hikaru.ps1
-			Start-Sleep -Seconds 5
-			shutdown -r -t 0
-			Start-Sleep -Seconds 30
-			exit
-        }
-    }
+$pendingreboot = (Test-Path 'HKLM:\SOFTWARE\Microsoft\Windows\CurrentVersion\Component Based Servicing\RebootPending')
+$pendingrebootcount = (Get-ItemProperty -Path "HKCU:\Software\AutoIDKU").PendingRebootCount
+if ($pendingreboot -eq $true -and $build -ge 10240) {
+	if ($pendingrebootcount -gt 3) {
+		Set-ItemProperty -Path "HKCU:\Software\AutoIDKU" -Name "HikaruMode" -Value 1 -Type DWord -Force
+		& $PSScriptRoot\hikaru.ps1
+		Write-Host -ForegroundColor Black -BackgroundColor Yellow "Your PC have queued a restart more than 3 times!" -n; Write-Host ([char]0xA0)
+		Write-Host -ForegroundColor Yellow "This is likely due to Windows Update being busy at the moment. I suggest checking the page in Settings for any on-going updates, or check in Task Manager for any WU-related processes and wait for them to finish if possible."
+		Write-Host -ForegroundColor White "If you wish to continue the script despite the pending restart, press Enter twice. Otherwise, please restart the system manually (the script will automatically resume when you do so)."
+		Read-Host
+		Read-Host
+		Set-ItemProperty -Path "HKLM:\SOFTWARE\Microsoft\Windows NT\CurrentVersion\Winlogon" -Name "Shell" -Value 'explorer.exe' -Type String
+		Set-ItemProperty -Path "HKCU:\Software\AutoIDKU" -Name "HikaruMode" -Value 0 -Type DWord -Force
+	} else {
+		Write-Host -ForegroundColor Cyan -BackgroundColor DarkGray "Your PC has a pending restart, which has to be done before running this script. Automatically restarting in 5 seconds..." -n; Write-Host ([char]0xA0)
+		Set-ItemProperty -Path "HKCU:\Software\AutoIDKU" -Name "HikaruMode" -Value 1 -Type DWord -Force
+		& $PSScriptRoot\hikaru.ps1
+		$pendingrebootcounting = $pendingrebootcount + 1
+		Set-ItemProperty -Path "HKCU:\Software\AutoIDKU" -Name "PendingRebootCount" -Value $pendingrebootcounting -Type DWord -Force
+		Start-Sleep -Seconds 5
+		shutdown -r -t 0
+		Start-Sleep -Seconds 30
+		exit
+	}
 }
+
 Import-Module BitsTransfer -Verbose
 
 
 ##### ------ Functions ------
 Write-Host -ForegroundColor Cyan -BackgroundColor DarkGray "Importing Functions" -n; Write-Host ([char]0xA0)
-function Stop-Script { # This stops the Music running in the PowerShell 7 process before exiting
-	Stop-Process -Name pwsh -Force
+$global:musicplayer = ""
+function Stop-Script {
+	Set-ItemProperty -Path "HKCU:\Software\AutoIDKU" -Name "HikaruMusicStop" -Value 1 -Type DWord -Force
+	Stop-Process -Name "FFPlay" -Force -ErrorAction SilentlyContinue
 	exit
 }
 
@@ -246,120 +183,121 @@ Take-KeyPermissions $rootKey $key $sid $recurse
 Write-Host " "
 $dotnet35done = (Get-ItemProperty -Path "HKCU:\Software\AutoIDKU").dotnet35
 $dotnet462done = (Get-ItemProperty -Path "HKCU:\Software\AutoIDKU").dotnetrebooted
+$setupmusic = (Get-ItemProperty -Path "HKCU:\Software\AutoIDKU").HikaruMusic
 $manualstuffs = $true # Please DO NOT set to false
+if ($setupmusic -eq 1) {
+	Write-Host "Setting up Music" -BackgroundColor DarkGray -ForegroundColor Cyan -n; Write-Host ([char]0xA0)
+	Start-Process powershell -Wait -ArgumentList "-Command $workdir\music\musicn.ps1" -WorkingDirectory $workdir\music
+	Start-Process powershell -WindowStyle Hidden -ArgumentList "-Command $workdir\music\musics.ps1"
+}
+
 
 switch ($true) {
 	
-    default {
-        Write-Host "You did not select anything to do and will be taken back to the configuration screen next time you start the script. Press Enter to exit." -ForegroundColor Red -BackgroundColor DarkGray -n; Write-Host ([char]0xA0)
+	default {
+		Write-Host "You did not select anything to do and will be taken back to the configuration screen next time you start the script. Press Enter to exit." -ForegroundColor Red -BackgroundColor DarkGray -n; Write-Host ([char]0xA0)
 		Set-ItemProperty -Path "HKCU:\Software\AutoIDKU" -Name "ConfigSet" -Value 0 -Type DWord -Force
 		Set-ItemProperty -Path "HKCU:\Software\AutoIDKU" -Name "RebootScript" -Value 0 -Type DWord -Force
-        Read-Host
+		Read-Host
 		exit
-    }
-
-    $setupmusic {
-		Write-Host "Setting up Music" -BackgroundColor DarkGray -ForegroundColor Cyan -n; Write-Host ([char]0xA0)
-		Start-Process powershell -Wait -ArgumentList "-Command $workdir\music\musicn.ps1" -WorkingDirectory $workdir\music
-        Start-Process pwsh -ArgumentList "-WindowStyle Hidden -Command $workdir\modules\essential\setupmusic.ps1"
-    }
+	}
 	
 	{$dotnet35 -eq $true -and $dotnet35done -ne 1} {
-        Write-Host "Enabling .NET 3.5" -ForegroundColor Cyan -BackgroundColor DarkGray -n; Write-Host ([char]0xA0)
-        Enable-WindowsOptionalFeature -Online -FeatureName "NetFx3"
+		Write-Host "Enabling .NET 3.5" -ForegroundColor Cyan -BackgroundColor DarkGray -n; Write-Host ([char]0xA0)
+		Enable-WindowsOptionalFeature -Online -FeatureName "NetFx3"
 		Set-ItemProperty -Path "HKCU:\Software\AutoIDKU" -Name "dotnet35" -Value 1 -Type DWord -Force
 	}
 	
 	{$dotnet462 -eq $true -and $dotnet462done -ne 1} {
 		& $workdir\modules\apps\dotnet462install.ps1
-    }
+	}
 	
 	$manualstuffs {
 		& $workdir\modules\essential\manualstuffs.ps1
 	}
 	
-    $hidetaskbaricons {
-        & $workdir\modules\taskbar\hidetaskbaricons.ps1
-    }
+	$hidetaskbaricons {
+		& $workdir\modules\taskbar\hidetaskbaricons.ps1
+	}
 
-    $removeonedrive {
-        & $workdir\modules\removal\removeonedrive.ps1
-    }  
+	$removeonedrive {
+		& $workdir\modules\removal\removeonedrive.ps1
+	}  
 
-    {$removehomegroup -eq $true -and $build -lt 17134} {
-        & $workdir\modules\removal\removehomegroup.ps1
-    }
+	{$removehomegroup -eq $true -and $build -lt 17134} {
+		& $workdir\modules\removal\removehomegroup.ps1
+	}
 
-    $desktopshortcuts {
-        & $workdir\modules\desktop\desktopshortcuts.ps1
-    }
+	$desktopshortcuts {
+		& $workdir\modules\desktop\desktopshortcuts.ps1
+	}
 
-    $essentialapps {
-        & $workdir\modules\apps\essentialapps.ps1
-    }
+	$essentialapps {
+		& $workdir\modules\apps\essentialapps.ps1
+	}
 
-    $removeedgeshortcut {
-        & $workdir\modules\removal\removeedgeshortcut.ps1
-    }
+	$openshellconfig {
+		& $workdir\modules\apps\openshellconfig.ps1
+	}
 
-    $removewaketimers {
-        & $workdir\modules\removal\removewaketimers.ps1
-    }
-    
-    $removeUWPapps {
+	$removewaketimers {
+		& $workdir\modules\removal\removewaketimers.ps1
+	}
+
+	$removeUWPapps {
 		Write-Host -ForegroundColor Cyan -BackgroundColor DarkGray "Removing all UWP apps possible" -n; Write-Host ([char]0xA0) 
-        Start-Process powershell -Wait -ArgumentList "$workdir\modules\removal\removeuwpapps.ps1"
-    }
+		Start-Process powershell -Wait -ArgumentList "$workdir\modules\removal\removeuwpapps.ps1"
+	}
 
-    $taskbarpins {
-        & $workdir\modules\taskbar\removetaskbarpinneditems.ps1
-    }
+	$taskbarpins {
+		& $workdir\modules\taskbar\removetaskbarpinneditems.ps1
+	}
 
-    {$explorericon -eq $true} {
-        & $workdir\modules\taskbar\explorericon.ps1
-    }
-	
-    $disableaddressbar {
-        & $workdir\modules\apps\addressbar.ps1
-    }
+	$explorericon {
+		& $workdir\modules\taskbar\explorericon.ps1
+	}
 
-    $oldbatteryflyout {
-        & $workdir\modules\taskbar\oldbatteryflyout.ps1
-    }
+	$disableaddressbar {
+		& $workdir\modules\apps\addressbar.ps1
+	}
 
-    $openshellconfig {
-        & $workdir\modules\apps\openshellconfig.ps1
-    }
+	$oldbatteryflyout {
+		& $workdir\modules\taskbar\oldbatteryflyout.ps1
+	}
 
-    $registrytweaks {
-        & $workdir\modules\essential\simpleregistry.ps1
-    }
+	$registrytweaks {
+		& $workdir\modules\essential\simpleregistry.ps1
+	}
 
-    $customsounds {
-        Write-Host -ForegroundColor Cyan -BackgroundColor DarkGray "Installing custom system sounds" -n; Write-Host ([char]0xA0) 
-        Start-Process powershell -Wait -ArgumentList "$workdir\modules\desktop\customsounds.ps1"
-    }
+	$customsounds {
+		Write-Host -ForegroundColor Cyan -BackgroundColor DarkGray "Installing custom system sounds" -n; Write-Host ([char]0xA0) 
+		Start-Process powershell -Wait -ArgumentList "$workdir\modules\desktop\customsounds.ps1"
+	}
 
-    $replaceemojifont {
-        & $workdir\modules\removal\replaceemojifont.ps1
-    }
-	
+	$replaceemojifont {
+		& $workdir\modules\removal\replaceemojifont.ps1
+	}
+
+	$removeedgeshortcut {
+		& $workdir\modules\removal\removeedgeshortcut.ps1
+	}
+
 	$removesystemapps {
 		Write-Host -ForegroundColor Cyan -BackgroundColor DarkGray "Disabling system apps" -n; Write-Host ([char]0xA0) 
-        Start-Process powershell -Wait -ArgumentList "$workdir\modules\removal\removesystemapps.ps1"
+		Start-Process powershell -Wait -ArgumentList "$workdir\modules\removal\removesystemapps.ps1"
 	}
-	
+
 	$sltoshutdownwall {
 		& $workdir\modules\desktop\slidetoshutdownwall.ps1
 	}
-	
+
 }
 
 Set-ItemProperty -Path "HKCU:\Software\AutoIDKU" -Name "RebootScript" -Value 0 -Type DWord -Force
 & $PSScriptRoot\hikarinstall.ps1
 Write-Host " " -n; Write-Host ([char]0xA0)
 Write-Host "This was the final step of the script. Press Enter to restart and complete the setup" -ForegroundColor Black -BackgroundColor Green -n; Write-Host ([char]0xA0)
-Start-Process "$PSScriptRoot\ambient\FFPlay.exe" -Wait -NoNewWindow -ArgumentList "-i $PSScriptRoot\ambient\DomainCompletedAll.mp3 -nodisp -hide_banner -autoexit -loglevel quiet"
+Start-Process "$PSScriptRoot\ambient\FFPlay.exe" -Wait -WindowStyle Hidden -ArgumentList "-i $PSScriptRoot\ambient\DomainCompletedAll.mp3 -nodisp -hide_banner -autoexit -loglevel quiet"
 & $PSScriptRoot\notefinish.ps1
 Write-Host " "; Show-Branding; Write-Host "Made by Bionic Butter with Love <3" -ForegroundColor Magenta -n; Write-Host ([char]0xA0)
 Read-Host
